Initial implementation
The initial implementation and basic functionality has been created.
Input is split by spaces, for each element in the input stream if it is
an integer it will be pushed onto the stack, and if it is an operand the
operation will be performed upon the last two values of the stack.