More weird parsing
So currently the function is 95% of the way there in terms of
reimplementing the weird behaviour of the sprn, including the parsing of
all of the strange inputs provided. The one thing that I haven't got
working properly is the evaluation of "x-+y" where x and y are numbers.
This calculation would otherwise be read as "x", "-", "y", "+" following
the other logic that appears to work, but the sprn we are provided with
reads this as "-(x+y)", which just seems bonkers and I'm not quite sure
how to implement it. Unless I take all of the values and put them on one
temp array, then take all of the operators. The values go in in the
correct order, and operators go in in reverse order?